move property .lastProgress into the Effect class
