+++ packages/flame/lib/src/effects/component_effect.dart
@@ -1,5 +1,3 @@
-import 'package:flutter/cupertino.dart';
-
 import '../../components.dart';
 import 'controllers/effect_controller.dart';
 import 'effect.dart';
@@ -12,11 +10,6 @@ abstract class ComponentEffect<T extends Component> extends Effect {
   ComponentEffect(EffectController controller) : super(controller);
 
   late T target;
-
-  /// The effect's `progress` variable as it was the last time that the
-  /// `apply()` method was called. Mostly used by the derived classes.
-  double get previousProgress => _lastProgress;
-  double _lastProgress = 0;
 
   @override
   void onMount() {
@@ -31,24 +24,5 @@ abstract class ComponentEffect<T extends Component> extends Effect {
     } else {
       throw UnsupportedError('Can only apply this effect to $T');
     }
-  }
-
-  // In the derived class, call `super.apply()` last.
-  @mustCallSuper
-  @override
-  void apply(double progress) {
-    _lastProgress = progress;
-  }
-
-  @override
-  void reset() {
-    super.reset();
-    _lastProgress = 0;
-  }
-
-  @override
-  void resetToEnd() {
-    super.resetToEnd();
-    _lastProgress = 1;
   }
 }
