+++ packages/flame/lib/src/effects/remove_effect.dart
@@ -11,5 +11,6 @@ class RemoveEffect extends Effect {
     if (progress == 1) {
       parent?.removeFromParent();
     }
+    super.apply(progress);
   }
 }
