+++ packages/flame/test/effects/effect_test.dart
@@ -13,6 +13,7 @@ class _MyEffect extends Effect {
   @override
   void apply(double progress) {
     x = progress;
+    super.apply(progress);
   }
 
   @override
